+++ lib/keymap.vi_digraph.sh
@@ -59,7 +59,7 @@ function ble-decode/keymap:vi_digraph/initialize {
 
   ble/edit/info/immediate-show text "ble.sh: updating cache/keymap.vi_digraph..."
 
-  : >| "$fname_keymap_cache"
+  >| "$fname_keymap_cache"
   ble/decode/keymap#load vi_digraph dump 3>> "$fname_keymap_cache"
 
   ble/edit/info/immediate-show text "ble.sh: updating cache/keymap.vi_digraph... done"