+++ docs/ChangeLog.md
@@ -186,6 +186,7 @@
 - history: prevent `SIGPIPE` from reverting the TTY state in trap `EXIT` `#D2153` 4b8a0799
   - history: fix initially shifted history index `#D2180` e425dc56
 - edit: support `bleopt internal_exec_int_trace` (motivated by tessus) `#D2171` cebea478 3801a87e
+- global: avoid using the builtin `:` `#D2192` xxxxxxxx
 
 <!---------------------------------------------------------------------------->
 # ble-0.4.0-devel3